refactor: move commands definition to cmd package for testability
but no tests at the time
hope to add tests in the future
